% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup.R
\name{compare_elim}
\alias{compare_elim}
\title{Performs model runs for two EIR values and returns whether they achieve elimination}
\usage{
compare_elim(EIR_min, EIR_max, r_EM0, em_loss, surv_bioassay,
  bites_Emanator, bites_Indoors, bites_Bed, em_cov, itn_cov, Q0)
}
\description{
Performs model runs for two EIR values and returns whether they achieve elimination
}
\author{
Joel Hellewell
}
\keyword{internal}
