% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_indoor}
\alias{dat_indoor}
\title{Data frame of indoor biting times}
\format{
An object of class \code{data.frame} with 24 rows and 14 columns.
}
\usage{
dat_indoor
}
\description{
Data frame of indoor biting times
}
\keyword{datasets}
