% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup.R
\name{elim}
\alias{elim}
\title{Returns whether a run has achieved elimination}
\usage{
elim(mod_run)
}
\value{
TRUE or FALSE depending on whether elimination is achieved
}
\description{
Returns whether a run has achieved elimination
}
\author{
Joel Hellewell
}
\keyword{internal}
