% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{figure4e}
\alias{figure4e}
\title{Plot Figure 4e}
\usage{
figure4e()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of outdoor exposure.
}
\author{
Joel Hellewell
}
