% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{figure4f}
\alias{figure4f}
\title{Plot Figure 4f}
\usage{
figure4f()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of outdoor exposure.
}
\author{
Joel Hellewell
}
