% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure5.R
\name{figure5d}
\alias{figure5d}
\title{Plot Figure 5d}
\usage{
figure5d()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of bioassay survival in the local arabiensis-like vector population.
}
\author{
Joel Hellewell
}
