% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure6.R
\name{figure6a}
\alias{figure6a}
\title{Plot Figure 6a}
\usage{
figure6a()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of bioassay survival in the local gambiae-like vector population.
}
\author{
Joel Hellewell
}
