% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure6.R
\name{figure6b}
\alias{figure6b}
\title{Plot Figure 6b}
\usage{
figure6b()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of bioassay survival in the local gambiae-like vector population. This plot includes a 20% mortality effect.
}
\author{
Joel Hellewell
}
