% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure6.R
\name{figure6c}
\alias{figure6c}
\title{Plot Figure 6c}
\usage{
figure6c()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of bioassay survival in the local gambiae-like vector population. This plot is for optimal emanators with 100% repellency and no decay in effect over time.
}
\author{
Joel Hellewell
}
