% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure6.R
\name{figure6d}
\alias{figure6d}
\title{Plot Figure 6d}
\usage{
figure6d()
}
\value{
plot
}
\description{
This performs a unknown amount of model runs. It systematically searches for the maximum pre-intervention EIR that ITN and ITN+emanator
combinations will eliminate for varying levels of bioassay survival in the local arabiensis-like vector population.
}
\author{
Joel Hellewell
}
