% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{figuresupp1}
\alias{figuresupp1}
\title{Plot Figure supp 1
This performs no model runs.}
\usage{
figuresupp1()
}
\value{
plot
}
\description{
Plot Figure supp 1
This performs no model runs.
}
\author{
Joel Hellewell
}
