% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup.R
\name{find_EIR_boundary}
\alias{find_EIR_boundary}
\title{Finds the EIR boundary systematically for a given step size}
\usage{
find_EIR_boundary(step, EIR_min, r_EM0, em_loss, surv_bioassay,
  bites_Emanator, bites_Indoors, bites_Bed, em_cov, itn_cov)
}
\value{
Minimum EIR value
}
\description{
Finds the EIR boundary systematically for a given step size
}
\author{
Joel Hellewell
}
\keyword{internal}
