% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_setup.R
\name{mr}
\alias{mr}
\title{Model run
This function performs model runs with specified parameters. Some default parameters are specified to save repeating code. Some must
be specified each time to guarantee that parameter values aren't incorrectly passed.}
\usage{
mr(
  em_cov,
  itn_cov,
  bites_Emanator,
  bites_Indoors,
  bites_Bed,
  r_EM0 = 0.6053263,
  em_loss = 0.001954954,
  d_EM0 = 0,
  Q0 = 0.92,
  surv_bioassay = 0,
  init_EIR,
  ...
)
}
\value{
A list of all of the different outputs from the model
}
\description{
Model run
This function performs model runs with specified parameters. Some default parameters are specified to save repeating code. Some must
be specified each time to guarantee that parameter values aren't incorrectly passed.
}
\author{
Joel Hellewell
}
